Make files for all assignments
