+++ index.html
@@ -2,9 +2,11 @@
 <html lang="en">
 <head>
     <meta charset="UTF-8">
+    <link rel="stylesheet" href="./styles.css">
     <title>Tech It Easy Dashboard</title>
 </head>
 <body>
 
+<script src="main.js" type="module"></script>
 </body>
 </html>
