+++ opdracht1.js
@@ -1,0 +1,21 @@
+import inventory from "./inventory.js";
+
+/*
+Opdracht 1a: Gebruik een array-methode om een array te maken met alle tv-type namen. Log de uitkomst in de console.
+
+    pseudo code
+
+Input:
+-Array
+
+Output:
+-Array met entries met tv type namen
+-Log in de console
+
+Stappen:
+-Loop door array
+-Selecteer bij iedere entry de tv type naam
+-Sla de tv type naam op
+-Maak van alle tv type namen een array
+*/
+
